# FoodTime_Backend/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FoodTime_Backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/GestionarMenus/all": {
      "get": {
        "tags": [
          "Menus"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/GestionarMenus/Crear": {
      "post": {
        "tags": [
          "Menus"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/GestionarMenus/{id}": {
      "get": {
        "tags": [
          "Menus"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Menus"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Menus"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ListarPedido": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Pedidos/{id}": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Pedidos/Crear": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/": {
      "put": {
        "tags": [
          "Pedidos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pedidos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "EstadoPedido": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4
      ],
      "type": "integer"
    },
    "LineaPedido": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "menu": {
          "$ref": "#/definitions/Menu"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Menu": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "titulo": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "tipoMenu": {
          "$ref": "#/definitions/TipoMenu"
        },
        "disponibilidad": {
          "type": "boolean"
        }
      }
    },
    "Pedido": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "$ref": "#/definitions/EstadoPedido"
        },
        "lineaPedidos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LineaPedido"
          }
        },
        "total": {
          "format": "double",
          "type": "number"
        },
        "tipoEntrega": {
          "$ref": "#/definitions/TipoEntrega"
        },
        "horaEntrega": {
          "type": "string"
        }
      }
    },
    "TipoEntrega": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    },
    "TipoMenu": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    }
  }
}