# FoodTime_Modelo/Modelo/TipoEntrega.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace FoodTime_Modelo.Modelo
{
    public enum TipoEntrega
    {
        INMEDIATO,
        HORAESPERADA
    }
}