# FoodTime_Service/DTOs/LineaPedidoDTO.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace FoodTime_Service.DTOs
{
    public class LineaPedidoDTO
    {
    }
}